+++ src/pyall/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "0.3.1"
+__description__ = "Pyall is a linter that tries to keep the __all __ in your Python modules always up to date."  # noqa: E501