+++ src/pyall/__main__.py
@@ -1,6 +1,4 @@
-import sys
-
 from pyall.main import main
 
 if __name__ == "__main__":
-    sys.exit(main())
+    raise SystemExit(main())