+++ src/pyall/constants.py
@@ -1,28 +1,13 @@
 import ast
-from typing import Any, Callable, TypeVar, Union
 
 __all__ = [
     "ADD_COMMENTS_REGEX_PATTERN",
     "ALL_NODE",
-    "ASTFunctionT",
-    "DESCRIPTION",
     "EXCLUDE_REGEX_PATTERN",
-    "Function",
     "GLOB_PATTERN",
     "INCLUDE_REGEX_PATTERN",
     "SKIP_COMMENTS_REGEX_PATTERN",
-    "VERSION",
 ]
-
-VERSION = "0.1.0"
-DESCRIPTION = (
-    "Pyall is a linter that tries to keep "
-    "the __all __ in your Python modules always up to date."
-)
-
-# TYPE
-Function = TypeVar("Function", bound=Callable[..., Any])
-ASTFunctionT = Union[ast.FunctionDef, ast.AsyncFunctionDef]
 
 # TUPLE
 ALL_NODE = (ast.ClassDef, ast.FunctionDef, ast.AsyncFunctionDef, ast.Name)