+++ docs/changelog.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.1] - 17/07/2022
+
+- Apply src layout
+- Support Python3.10
+- Add py.typed
+
+## [0.2.0] - 11/04/2021
+
+- [Version 0.2.0 by @hakancelik96](https://github.com/hakancelikdev/pyall/pull/5)
+  - Docs update
+  - pyall: public ( bug fix )
+
 ## [0.1.0] - 11/04/2021
 
 - [Version V0.1.0 by @hakancelikdev](https://github.com/hakancelikdev/pyall/pull/1)